Fix symbol->string and string->symbol.
The former did not correctly determine the length of the symbol name,
resulting in "" all the time. The latter did not properly intern the new
symbol (in case it was not known yet), resulting in

    (equal? (string->symbol "newfoo") (string->symbol "newfoo"))

to return #f, and in consequence resulting in the same module being
loaded multiple times.